Added basic functionality to get temp, humidity and status
Basically just made the ability to get temp and humidity from the SENSE HAT.  Open the config,json, read the value of min_humidity and return "OK" if the humidity from the sense hat is BELOW the one specified in the config.json file. Please note THIS IS NOT OOP, this is all procedural code, OOP and better programming semantics coming soon :) Both myself and Sherry worked on this commit together.